+++ src/json-schema.js
@@ -87,7 +87,7 @@ function convertSubSchema(schema, references) {
   return actualSchema;
 }
 
-function lookupReference(reference, root) {
+export function parseReference(reference) {
   const parts = reference.split('/');
 
   if (parts[0] !== '#') {
@@ -99,6 +99,12 @@ function lookupReference(reference, root) {
   }
 
   const id = parts[2];
+
+  return id;
+}
+
+function lookupReference(reference, root) {
+  const id = parseReference(reference);
 
   if (!root.definitions || !root.definitions[id]) {
     throw new Error(`Reference to ${reference} does not exist`);
@@ -130,19 +136,21 @@ function checkSchemaHasReferences(schema) {
 
 /** Convert Swagger schema to JSON Schema
  */
-export default function convertSchema(schema, root) {
+export default function convertSchema(schema, root, convertDefinitions = true) {
   const references = [];
   const result = convertSubSchema(schema, references);
 
-  if (references.length !== 0) {
-    result.definitions = {};
-  }
+  if (convertDefinitions) {
+    if (references.length !== 0) {
+      result.definitions = {};
+    }
 
-  while (references.length !== 0) {
-    const lookup = lookupReference(references.pop(), root);
+    while (references.length !== 0) {
+      const lookup = lookupReference(references.pop(), root);
 
-    if (result.definitions[lookup.id] === undefined) {
-      result.definitions[lookup.id] = convertSubSchema(lookup.referenced, references);
+      if (result.definitions[lookup.id] === undefined) {
+        result.definitions[lookup.id] = convertSubSchema(lookup.referenced, references);
+      }
     }
   }